+++ nyanPictureLib/blt256Beta.cpp
@@ -28,8 +28,8 @@ void CBlt256Beta::Print(POINT putPoint,POINT srcPoint,SIZE putSize,LPVOID picDat
 	int screenSizeX = CMyGraphics::GetScreenSizeX();
 //	int screenSizeY = CMyGraphics::GetScreenSizeY();
 
-	src += srcPoint.y * srcSize.cx + srcPoint.x;
-	dst += putPoint.y * screenSizeX + putPoint.x;
+	src += (SSIZE_T)srcPoint.y * srcSize.cx + srcPoint.x;
+	dst += (SSIZE_T)putPoint.y * screenSizeX + putPoint.x;
 
 	int srcPitch = srcSize.cx * sizeof(char);
 	int dstPitch = screenSizeX * 4;
